Fix mkdtemp() umask issue
Even though the mkdtemp() man page indicates that the
directory will be created with always 0700 permissions,
it is actually affected by the process's umask.

So, if you run stap with an unusual umask it can end
up creating the temp dir with permissions that
staprun can't handle e.g.:

  $> rpm -q systemtap
  systemtap-0.6-1.fc9
  $> umask 0122
  $> stap -e 'probe begin { println("foo") exit() }'
  ERROR: Error opening '/tmp/stapV4pBIb/stap_725b9bc541cef2618a5ccbc58bb64d15_287.ko': Permission denied

Obvious solution is to briefly set the umask to zero
in stap so as to ensure 0700 permissions.

Signed-off-by: Mark McLoughlin <markmc@redhat.com>