Add test suite for declaration resolution
This adds the test declaration.exp which checks that specific named
declarations inside the kernel both fail and succeed.  We get the
failure by only using a single compile unit which is currently known
to have the structure stubbed.  We get exactly the same declaration to
succeed by adding a CU that is known to have the declaration.

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>