Add hack to support git 1.6 in git_version.sh
This is needed now for people who are running the "master" branch of
git, which is currently in development for git version 1.6.0.

This is a kludge to support hyphenated git commands (i.e.,
git-diff-files), but this form is explicitly deprecated, and doesn't
work in git 1.6 unless you put git's execdir (obtained via git
--exec-path) in your PATH environment variable.  Even then, these
programs may no longer exist future versions of git.  It would be
better to change the script to use "git diff-files", et. al., but I
assume upstream had some vaguely justifiable reason for not writing
the script that way.  I'll let them decide when they are willing to
fix the script the right way, even if it means breaking compatibility
for some ancient version of git.  n.b., git 1.5.0 was released
February 2007, and git 1.4.0 was released June, 2006.

Signed-off-by: "Theodore Ts'o" <tytso@mit.edu>