fix missing functions in caches
Both the inline and non inline filtered function caches can lose
functions.  This happens because both of these caches are indexed by the
entrypc.  However, the vagaries of optimisation can actually cause us to
end up with two functions with apparently the same entrypc (mainly
because for inline functions, the entrypc has to be deduced from the
range of addresses, which is nastily heuristic).

A nice example of this is bsg_io_schedule.  In my build of 2.6.26 it's a
static function transformed to an inline by the compiler.  The range
deduction of its entrypc ends up being identical to that of bsg_release,
so when I try to place a probe inside it, none of the function
references actually refers to it.

Since nothing ever actually makes use of this index, the fix is simple:
don't use a map indexed on the entrypc, use a simple vector.

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>