Update guidelines for the ChangeLog-less world
Getting rid of ChangeLogs doesn't mean that we get a free ticket -- we
now need to be more diligent about providing meaningful commit messages.
I've updated the HACKING file with a first draft of new guidelines, but
we may still revise what we feel is appropriate detail in the logs.

I removed the ChangeLog section from the tapset/DEVGIDE entirely, and
also fixed the path where examples are stored.