PR9932: use @cast module search path
The nfs, rpc, and scsi tapsets use @cast on types that may be compiled
into a kernel module or into the main kernel binary.  The @cast search
path separated with colons lets us search both the kernel and the module.

For a couple of cases, I also merged sequential @casts that work just
fine as a single cast with a multiple-level dereference.