Move lookup_bad_addr call in STAPCONF_PROBE_KERNEL
With most of the implementations, kread/kwrite call deref/store_deref,
and so it makes sense to have lookup_bad_addr in the latter as an
underlying address check.

However, in the STAPCONF_PROBE_KERNEL case that uses probe_kernel_read
and probe_kernel_write, the roles are reversed, so lookup_bad_addr needs
to be in kread/kwrite.

Also note that __deref_bad and __store_deref_bad should only be used in
cases that can be determined at compile time.  These turn into invalid
symbols which prevent the module from loading.  (They might be better
replaced with compile-time assertions.)