Fix regression in tracepoint unregistration
Commit 96b030fe reorganized the tracepoint registration calls by
creating generic wrappers that return int.  However, the older
tracepoint implementation (as found in RHEL5.3) returned void for unreg,
so this was failing pass-4.

Since we can't handle unregistration failures anyway, this change just
makes the generic unregister function return void instead.  As noted in
the newly-added comment, it should be safe for us to ignore unreg
failures.