PR9951: Prevent GCC warnings in deref()
In some configurations, GCC was warning about a possible use of _v in
the deref macros.  I could not reproduce the error, but the only case
where _v is not written is if lookup_bad_addr rejects the address, in
which case we will hit DEREF_FAULT and _v won't be used.

Now we're priming _v=0 anyway, so GCC has no right to complain...