Add more file stats to the hash
For kernel developers, it may be common practice to reuse the same
kernel build tree for several kernel variants.  Our previous hashing
only considered the release version, architecture, and build path, which
may all remain constant for such a developer.

This change adds the file size and mtime of several kernel version files
to the hash, so it should be a bit more robust against collisions.