Fix runtime/itrace.c to call arch_has_*_step() prior to calling utrace_control
As Roland pointed out in his Mar 31 reply to a posting of mine (subject:
Re: Testing insn.block probe point uncovers possible utrace bug),
utrace_control() documents that the caller must ensure that
arch_has_single_step and arch_has_block_step are defined before trying
to use those step modes.  The attached patch addresses that issue.  I've
tested this patch on x86_64 arch, and a block step test runs
successfully, since block step is supported on that arch.  Testing on
ppc64 arch, the test fails as expected (since block step is not
supported on ppc64 yet) with:
	"ERROR: callback for <pid> failed: 1"
which is sent to stdderr and
	"usr_itrace_init: arch does not support block step mode"
which is sent to the system log.

This isn't the most user-friendly way of surfacing an error.  Perhaps
the stap runtime could have a set of defined return codes that would be
mapped to strings so the user can get an idea of what the error is
without looking in the system log.  But that's a side issue, of course.

Regards,
-Maynard