PR10016: Purge stap of all pgrp and system() usage
We hereby no longer try to manipulate process groups in any way.  We
don't set a private process group, and we never kill() our entire group
either.  Instead of using system(), we now have a stap_system() which
saves the child PID, so when we get a terminating signal we can pass it
along to the child.

Signals sent through the TTY have always worked, since the TTY sends it
to the entire pgrp.  However, if we're running as part of a wrapper
script or GUI, which may not have a separate process group for stap, we
still would like to allow "kill -TERM $STAPPID" to terminate stap
nicely.

There's still a short window of failure in the time that staprun is
active, because we can't kill a setuid process from a user process.
Once staprun drops privileges and execs to stapio though, everything
should work fine.