PR10026: Read marker/tracepoint args directly
We already stash the context variables for markers and tracepoints into
the locals for the probe body, but then we were using separate functions
to read those locals for each particular probe body.

This patch instead teaches the unparser how to emit the local name
directly for those context variables.  The resulting code from the
translator is much simpler now.