PR10049: Restore better errors for $return
In commit c4ce66a1 I shifted some of the variable accesses around, and
that caused the error messages for $return variables to lose context.

The root problem in this case was that dwarf_attr didn't find the type
attribute, and that failure wasn't immediately caught.  Now that result
is checked so we can provide a proper error message.

While I'm at it, changing that lookup to dwarf_attr_integrate is enough
to make the $return value usable again, at least for the reported test
case.