syscalls testsuite: use different escape sequence
The syscall testsuite uses embedded "// TEXT" markers in the .c files
to designate expected output, kind of like the "dg-" bunch in the gcc
test suite.  Unfortunately, "//" is not a unique prefix to systemtap
tests, and in particular it can occur in the system headers that will
be picked up with the "gcc -E -C ..." invocation in test.tcl.

So let's switch to "//staptest//".  test.tcl is also modified to
escape a few more mischevious regexp metacharacters that might sneak
past.