Simplify buildok/context_test.stp
The test was using optional probes on uptime_read_proc, which doesn't
exist anymore on 2.6.30.  The problem is that when those optional probes
are skipped, the test doesn't really do anything.

For a buildok test of the context functions, the actual probe point
doesn't matter, so I've changed it to just a begin probe that calls all
of the functions.