Refactor anonymous lookups in translate_components
The previous code recursed the entirety of translate_components, and
it seemed to be restarting the components list every time, so it would
only work if the anonymous portion was the first component.  Even then,
examining the code output by semok/thirtythree revealed that it wasn't
fully translating the locations when multiple anonymous pieces were
involved.

Instead, it now recurses in a separate function, find_struct_member,
which does just enough to find the member die and return.  It also
builds a vector of the locations passed through, so translate_components
can output code for the full chain of anonymity.

The generated code for semok/thirtythree's $page->mapping now appears to
match the offsets from my manual inspection of struct page.  I also
added a test for $page->first_page->mapping, which works now but would
segfault the old code.