Fix another kernel/kprobe.function conflict
Both kernel.function and kprobe.function were defining a global array
stap_unreg_kprobes to use in bulk kprobes unregistration.  The compiler
allowed the duplicate definition as long as they were the same size, as
it was when exercised in buildok/thirtyone.

kprobe.function now uses a separate stap_unreg_kprobes2, and the
testcase is modified to produce an imbalanced number of probes.