Cache the last result of dwarf_getscopes
This one function accounted for ~30% of my callgrind profile of
"stap -l 'syscall.*'", even though it was only called ~1200 times.  We
call dwarf_getscopes for each $target variable, with the same parameters
within a given probe.  Since they're no nicely grouped, it's easy to
just cache the most recent call, and the next few calls will be a hit.

Overall this cuts the number of calls down to about 300, for an easy
speed gain.