Fix condition propagation across aliases
When an instance of an alias has a condition, that condition gets
propagated to each of the locations that the alias defines.  However,
the copy of the location list was not a deep copy, and so all other
instances of the alias would also incorrectly receive the condition.

This patch makes the location list copy a little deeper, and adds a
test case which demonstrates the issue.