PR10274 Fix exelib -O3 testcase.
The testcase now uses noinline, an empty asm statement and a volatile
variable to prevent the function getting inlined, being totally unrolled
or the recursive call being replaced with an inner-iteration.

This does defeat part of the testcase though, which was testing unwinding
through an optimized recursive function. But it seems the best we can do.

* testsuite/systemtap.exelib/exelib.exp: Add -O3 to the mix.
* testsuite/systemtap.exelib/uprobes_exe.c: Use volatile bar and mark
  main_func noinline.
* testsuite/systemtap.exelib/uprobes_lib.c: Use volatile foo and mark
  lib_func noinline.