Add str_replace() To Tapsets
This patch adds a search and replace string functionality to existing
tapsets.

The functionality is as follows:

The function takes in a parent string and searches for a substring as
specified by the user. If substring not found, the parent string is
returned. If substring is found, it is replaced by another string and
returned.

NOTE: The function will search and replace all the occurrence of
substrings in a parent string when matched.

Signed-off-by: Varun Chandramohan <varunc@linux.vnet.ibm.com>
Signed-off-by: Josh Stone <jistone@redhat.com>