Add seperate debuginfo after prelinking to exelib.exp tests.
We split debuginfo before prelinking, that means that the base load address
of the .so is non-zero, while for the .debug files they are zero.  In the
other case we supported sep-debug without prelinking both .so and .debug
load bases were zero. In this new test case both base load addresses are
non-zero.

* testsuite/systemtap.exelib/exelib.exp: Add libdebug == "sep-after" to
  indicate separate debuginfo after prelinking library. Skip sep-after
  if not prelinking.