PR10533 inlined vars are not always found (dwflpp).
dwflpp::find_variable_and_frame_base switched too early to "pyshical view"
of die tree. We need to lookup the var in the "syntactical view" of the die
tree first. Then when we see a DW_TAG_inlined_subroutine while walking the
syntactical die tree as returned by dwarf_getscopes for a given variable
to retrieve the frame base, then we need to switch to walking the physical
die tree where the subroutine is inlined.

* dwflpp.cxx (find_variable_and_frame_base): Don't immediately go to
  die_scopes, but lookup var first based on pc. Then switch "die branches"
  when searching for frame base and a DW_TAG_inlined_subroutine is
  encountered.