PR10538: Improve location lookup for unions
We had a bug that the starting call to find_struct_member used the same
memory for the parentdie and the resulting member.  If parentdie is a
union, then the first member probably won't have a location, and we
actually assert that it must be a union.  Since we wrote the result in
the same memory, we lost the real info about the parent, and so the
assertion failed.

* dwflpp.cxx (dwflpp::translate_components): Use distinct memory for the
  parent and resulting member in the call to find_struct_member.
  (dwflpp::find_struct_member): Remove the needless parentdie copy.