PR10537 process().function().label() should select multiple inlined instances.
This is less useful than one would hope. gcc will often emit a label with
a DW_AT_low_pc that is not really in the neighbourhood of where one would
expect it when the label is inlined and gcc can proof the label isn't really
used in the optimized code. dwflpp::iterate_over_labels will now really
iterate recursively through the die, even for dies without a name (like
lexical blocks). This means we should now always find the concrete inlined
label instances that have a real DW_AT_low_pc and so we don't need the trick
to use the line table to get at the actual address.

* dwflpp.cxx (iterate_over_labels): Accept dies without a name. Don't handle
  labels without a name or without a lowpc attribute.
* testsuite/systemtap.base/inlinedvars.c (m): Trick gcc into thinking label
  is always used.
  (call, call2): Activate.
  (main): Call call and call2.
* testsuite/systemtap.base/inlinedvars.exp: New result_string.
  Test both unoptimized and optimized (inlined) builds.