Cache inline instance lookups
We used to call dwarf_func_inline_instances to get the locations where
inlines are used.  This function has to iterate through nearly all DIEs
to find instances, which is a lot of redundant work when we're probing
multiple inline functions.

Now we have our own dwarf iterator to cache all inline instances back to
their origin.  This only needs to be called once for each CU, and all
further inlines are just a map lookup.  Some quick benchmarks:

  stap -l                           Before      After
  kernel.function("*")             25010ms     2110ms
  module("*").function("*")        86550ms    16920ms
  process("stap").function("*")    41330ms      580ms

* dwflpp.cxx (dwflpp::cu_inl_function_caching_callback): Removed.
  (dwflpp::cache_inline_instances): New caching iterator.
  (dwflpp::iterate_over_inline_instances): Cache each CU once.