Try to build tracequery for all headers at once
To mitigate PR10424, we switched to building a separate tracequery
module for each tracepoint header, so a bad header wouldn't break all of
the others.  However, with recent kernels that leads to ~18 make
commands, which adds up quickly in time.  It's cached, so that's not too
bad, but as a developer who rebuilds stap frequently, it gets annoying.

If we're going to call 18 makes, it's worth it to start with one bigger
make that covers all the headers at once (like we used to).  If that one
fails, we can still fall back to compiling individually.

FWIW, the failing ext4.h header was only created in 2.6.31, and was
fixed before 2.6.32, so the specific failure in PR10424 has a fairly
small window.

* buildrun.cxx (make_tracequery): Just take a single vector of headers.
* hash.cxx (find_tracequery_hash): Deal with multiple headers.
* tapsets.cxx (tracepoint_builder::get_tracequery_module): Ditto.
  (tracepoint_builder::init_dw): Attempt all system headers together,
  and if that fails, try again individually.