PR10678 vta-gcc: module debuginfo: relocation refers to undefined symbol
libdwfl tries to resolve all relocations in a module debuginfo file and
if it cannot find a symbol used in a relocation it will fail when
dwfl_module_getdwarf() is called. So we must make sure all possible
dependencies of the module are also in the dwfl. We do this by trying
to find and parse the modules.dep file and insert all dependencies
into the dwfl.

* setupdwfl.cxx (elfutils_kernel_path): Lift from setup_dwfl_kernel and
  make static.
  (is_comma_dash): New function.
  (modname_from_path): Likewise.
  (setup_mod_deps): Likewise.
  (setup_dwfl_report_kernel_p): Call setup_mod_deps().
* testsuite/buildok/pr10678.stp: New test.