Fix transok/tval-opt.stp testcase. Pick diffent function and non-empty block.
This testcase succeeded just because the value being set couldn't be
found. So the error message being compared was the same. Set -o pipefail
to catch that case. On vta compiled kernels it failed because the optimizer
turned { statement } into statement. So pick a function and argument which
location can always be found and add an extra 'next' statement so the
block isn't folded.

* testsuite/transok/tval-opt.stp: Set -o pipefail. Add 'next' to make
  sure block isn't empty. Use "do_filp_open" and "mode".