Explain "cheat" comment in sym.c (_stp_tf_mmap_cb).
We are abusing the "first" section address here to indicate where the
module (actually first segment) is loaded (which is why we are ignoring
the offset). It would be good to redesign the stp_module/stp_section
data structures to better align with the actual memory mappings we are
interested in (especially the "section" naming is slightly confusing
since what we really seem to mean are elf segments (which can contain
multiple elf sections).

* runtime/sym.c (_stp_tf_mmap_cb): Add cheat comment.