Fixup some memory tapset vm kernel function probe fallbacks.
Older kernels don't have all GFP constants defined, and the fallback
kernel function probe fallbacks don't have the same dwarf variable
names as the kernel trace point probes. So replace them with variables
that are available. bytes_alloc was sometimes a long and sometimes a
string, this caused scripts to fail depending on which alternative was
chosen for a particular kernel. So make it a long always.
This isn't a full solution since kmalloc is always inlined which makes
the kernel.function("kmalloc").return probe fail.

* tapset/memory.stp: Define __GFP_THISNODE, __GFP_RECLAIMABLE,
  GFP_TEMPORARY, GFP_HIGHUSER_MOVABLE and GFP_THISNODE when not yet
  defined.
  (__vm.kmalloc.kp): Use $flags, not $gfp_flags. Set bytes_alloc equal
  to bytes_req.
  (__vm.kmem_cache_alloc.kp): Likewise. And use $cachep->buffer_size
  for bytes_req.
  (__vm.kmalloc_node.kp): Likewise.
  (__vm.kmem_cache_alloc_node.kp): Likewise.
  (__vm.kfree.kp): Use $ibjp for ptr, not $return.
  (__vm.kmem_cache_free.kp): Likewise.
* testsuite/buildok/vm.tracepoints.stp: Move vm.kmalloc test to...
* testsuite/buildok/vm.tracepoints.kmalloc.stp: ... here.