Fix the listing mode of process.library.mark
Our hack for producing nicer listing modes was to write the mark name in
component index 1.  That's fine for process.mark, but writes the wrong
position for process.library.mark.  It now munges the last probe
component of any number of indexes.

(A better fix is due in PR10831, preserving the derivation chain.)