PR909: Runtime for Performance Event Sampling
Implements a very simple sampling runtime to using the performance
events kernel API. An perf event attribute describing the setup and a
function to handle the counter overflows are passed into
_stp_perf_init(). This function sets up the event on each processor.
If successfully initialized, a pointer data structure is
returned. When the sampling is no longer needed _stp_perf_del() is
called to shutdown the sampling.

 * runtime/perf.h: Add declarations for data structures and functions
 * runtime/perf.c: Remove old perfmon runtime runtime.
   		   Add  _stp_perf_init() and _stp_perf_del() functions.