Prevent accidental global cleanup triggering for syscall test.tcl.
Later tests might trigger a cleanup and might set the global dir variable
leading to the syscall test.tcl trying to do an exec rm -rf on whatever
dir was set. So rename proc cleanup and global dir in syscall test.tcl to
something a bit less likely to clash.

* testsuite/systemtap.syscall/test.tcl: Rename proc cleanup to syscall_cleanup
 and global dir to syscall_dir.
* testsuite/systemtap.syscall/test-debug.tcl: Likewise.