The rpc buildok test suite probe breaks with new kernels
With newer kernels the rpc_release_task() routine is not defined
because it becomes inline since its only called once. Plus the
task allocation routines have changed so the arguments no longer
match up with previous routines. So this patch uses the kernel_v
clauses to allow the original functionality with older kernels
and not to break on new kernels.

Signed-off-by: Steve Dickson <steved@redhat.com>