PR11346: Move the skip-badvars logic into const-folding
The const_folder is where we prune constant expressions, like
conditionals based on a @defined($foo), so we want to give this a chance
to work before skip-badvars comes in and throws a warning.

As an added bonus, this makes skip-badvars more generic, so it will work
even for non-dwarf $target variables.

* elaborate.cxx (const_folder::visit_target_symbol): Enact skip-badvars.
* tapsets.cxx (dwarf_var_expanding_visitor::visit_target_symbol): Don't
  worry about badvars here anymore.  Save the error now, clean up later.
* testsuite/semok/badvar_undefined.stp: Test that @defined with
  skip-badvars doesn't throw any warnings.