PR11473: Optimize references to iteration values
When iterating with "foreach", we keep a map_node* which tells us our
current place in the map (and there's a similar marker for @hist).  Most
of the time when users index a map within a foreach, it is to get the
current value using the keys given by foreach.  When we know this is the
case, we can grab the value from the iterator instead of again searching
the map.

In the example given in PR11472, with two stat-map accesses within a
foreach, I recorded 12-15% improvement in runtime.

* translate.cxx (aggvar::get_hist): New accessor for indexing a @hist.
  (mapvar::get_value): New accessor for reading a map iterator.
  (mapvar::get_key): Simplify string NULL-check as done in get_value.
  (c_unparser::visit_foreach_loop_value): New function to see if the
  keys of a foreach are constant in the loop, and thus whether
  references with those keys can be optimized.
  (c_unparser::get_foreach_loop_value): Check for an active iteration
  value which matches the given arrayindex.
  (c_tmpcounter::visit_foreach_loop): Call visit_foreach_loop_value.
  (c_unparser::visit_foreach_loop): Ditto.
  (c_unparser::load_aggregate): Try to use a foreach value.
  (c_unparser::visit_arrayindex): Ditto.
  (c_tmpcounter::visit_arrayindex): Skip some tmps when foreach has it.
  (c_tmpcounter::visit_print_format): Ditto.
  (c_tmpcounter::visit_stat_op): Ditto.