Let hist buckets be printed with other arguments
We already have quite a kludge to fall back from the print(@hist(...))
special case if the @hist happens to have a bucket arrayindex on it.  In
that case we're printing scalars, so we can allow the print to contain
other scalar values too.

As noted in a comment, this still fails if the arrayindex is the start
of a larger expression. :(

* parse.cxx (parser::parse_symbol): Parse other print args after @hist.
* testsuite/buildok/print_histogram_entry.stp: Add a few prints to
  exercise this (and one that still doesn't work).