tracequery build fix
Similar to commit 392e08b7 -- we need to #undef TRACE_INCLUDE_FILE
before each header in the tracepoint query module, so they can be
queried all at once.

The side effect was that the all-inclusive tracequery build would fail,
so we'd fall back to building a separate tracequery for each header.
It's much faster when we can build them together.

* buildrun.cxx (make_tracequery): #undef TRACE_INCLUDE_FILE before each
  tracepoint header to prevent macro redefinition.